some optimizations

+++ src/main/java/be/Balor/Manager/Commands/Player/Fireball.java
@@ -47,7 +47,7 @@ public class Fireball extends ACCommands {
 	@Override
 	public void execute(CommandSender sender, String... args) {
 		Player player = null;
-		float power = (Float)ACHelper.getInstance().getConfValue("DefaultFireBallPower");
+		double power = (Double)ACHelper.getInstance().getConfValue("DefaultFireBallPower");
 		if (args.length >= 1) {
 			try {
 				player = Utils.getUser(sender, args, permNode, 1, false);
